Reduce duplication in DNSTAP source
